+++ Particle/Particle/Sources/Main/SearchTab/SearchViewController.swift
@@ -32,100 +32,9 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
     
     var disposeBag = DisposeBag()
     
-    private enum Metric {
-        enum SearchBar {
-            static let topMargin = 24
-            static let horizontalMargin = 20
-            static let height = 48
-        }
-        
-        enum ListTitle {
-            static let topMargin = 32
-            static let leftMargin = 20
-        }
-        
-        enum RemoveButton {
-            static let rightMargin = 8
-        }
-        
-        enum List {
-            static let horizontalMargin = 20
-            static let topMargin = 12
-        }
-        
-        enum TagTitle {
-            static let topMargin = 65
-            static let leftMargin = 20
-        }
-        
-        enum Tags {
-            static let topMagin: CGFloat = 12
-            static let horizontalMargin: CGFloat = 20
-            static let minimumLineSpacing: CGFloat = 10
-            static let minimumInterItemSpacing: CGFloat = 10
-        }
+    private var mainView: SearchMainView {
+        return self.view as! SearchMainView
     }
-    
-    private let searchBar: UISearchBar = {
-        let searchBar = UISearchBar()
-        searchBar.backgroundColor = .clear
-        searchBar.tintColor = .clear
-        searchBar.placeholder = "검색어를 입력해 주세요."
-        searchBar.searchTextField.font = .particleFont.generate(style: .pretendard_Regular, size: 16)
-        return searchBar
-    }()
-    
-    private let recentSearchListTitleLabel: UILabel = {
-        let label = UILabel()
-        label.text = "최근 검색어"
-        label.setParticleFont(.y_headline, color: .particleColor.gray04)
-        return label
-    }()
-    
-    private let recentSearchListRemoveButton: UIButton = {
-        let button = UIButton()
-        var configuration = UIButton.Configuration.plain()
-        configuration.contentInsets = .init(top: 10, leading: 12, bottom: 10, trailing: 12)
-        configuration.attributedTitle = AttributedString("모두 지우기", attributes: AttributeContainer.init([
-            .font: UIFont.particleFont.generate(style: .pretendard_Regular, size: 12) ?? UIFont.systemFont(ofSize: 12),
-            .foregroundColor: UIColor.particleColor.gray03
-        ]))
-        
-        button.configuration = configuration
-        return button
-    }()
-    
-    private let recentSearchList: UITableView = {
-        let tableView = UITableView()
-        tableView.register(SearchListCell.self)
-        tableView.backgroundColor = .clear
-        return tableView
-    }()
-    
-    private let tagTitleLabel: UILabel = {
-        let label = UILabel()
-        label.text = "관심 태그로 검색"
-        label.setParticleFont(.y_headline, color: .particleColor.gray04)
-        return label
-    }()
-    
-    private let tagCollectionView: UICollectionView = {
-        let layout = LeftAlignedCollectionViewFlowLayout()
-        layout.estimatedItemSize = CGSize(width: 1, height: 1)
-        layout.minimumLineSpacing = Metric.Tags.minimumLineSpacing
-        layout.minimumInteritemSpacing = Metric.Tags.minimumInterItemSpacing
-        
-        let collectionView = DynamicHeightCollectionView(
-            frame: .zero,
-            collectionViewLayout: layout
-        )
-        collectionView.register(LeftAlignedCollectionViewCell.self)
-        collectionView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-        collectionView.backgroundColor = .clear
-        collectionView.showsVerticalScrollIndicator = false
-        
-        return collectionView
-    }()
     
     init() {
         super.init(nibName: nil, bundle: nil)
@@ -137,11 +46,13 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
         fatalError("init(coder:) has not been implemented")
     }
     
+    override func loadView() {
+        self.view = mainView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         setupInitialView()
-        addSubviews()
-        setConstraints()
         
         Observable.of([
         "최근 검색어어어어",
@@ -150,7 +61,7 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
         "최근 검색어어어어",
         "최근 검색어어어어"
         ])
-        .bind(to: recentSearchList.rx.items(
+        .bind(to: mainView.recentSearchList.rx.items(
             cellIdentifier: SearchListCell.defaultReuseIdentifier,
             cellType: SearchListCell.self)
         ) { tableView, item, cell in
@@ -159,7 +70,7 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
         .disposed(by: disposeBag)
         
         Observable.of(tags)
-        .bind(to: tagCollectionView.rx.items(
+            .bind(to: mainView.tagCollectionView.rx.items(
             cellIdentifier: LeftAlignedCollectionViewCell.defaultReuseIdentifier,
             cellType: LeftAlignedCollectionViewCell.self
         )) { collectionView, item, cell in
@@ -167,11 +78,11 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
         }
             .disposed(by: disposeBag)
         
-        tagCollectionView.rx.itemSelected
+        mainView.tagCollectionView.rx.itemSelected
             .subscribe { [weak self] indexPath in
             guard let self = self else { return }
             guard let index = indexPath.element else { return }
-            guard let selectedCell = self.tagCollectionView.cellForItem(at: index) as? LeftAlignedCollectionViewCell else {
+                guard let selectedCell = self.mainView.tagCollectionView.cellForItem(at: index) as? LeftAlignedCollectionViewCell else {
                 return
             }
             selectedCell.setSelected()
@@ -181,68 +92,8 @@ final class SearchViewController: UIViewController, SearchPresentable, SearchVie
     }
     
     private func setupInitialView() {
-        view.backgroundColor = .particleColor.black
+        mainView.backgroundColor = .particleColor.black
         navigationController?.isNavigationBarHidden = true
-    }
-}
-
-// MARK: - Layout Setting
-
-private extension SearchViewController {
-    func addSubviews() {
-        [searchBar]
-               .forEach {
-                   self.view.addSubview($0)
-               }
-        
-        [
-            recentSearchListTitleLabel,
-            recentSearchListRemoveButton,
-            recentSearchList
-        ]
-            .forEach {
-                self.view.addSubview($0)
-            }
-        
-        [tagTitleLabel, tagCollectionView]
-            .forEach {
-                self.view.addSubview($0)
-            }
-       }
-    
-    func setConstraints() {
-        searchBar.snp.makeConstraints { make in
-            make.top.equalTo(self.view.safeAreaLayoutGuide).inset(Metric.SearchBar.topMargin)
-            make.left.right.equalToSuperview().inset(Metric.SearchBar.horizontalMargin)
-            make.height.equalTo(Metric.SearchBar.height)
-        }
-        
-        recentSearchListTitleLabel.snp.makeConstraints { make in
-            make.top.equalTo(searchBar.snp.bottom).offset(Metric.ListTitle.topMargin)
-            make.left.equalToSuperview().inset(Metric.ListTitle.leftMargin)
-        }
-        
-        recentSearchListRemoveButton.snp.makeConstraints { make in
-            make.centerY.equalTo(recentSearchListTitleLabel)
-            make.right.equalToSuperview().inset(Metric.RemoveButton.rightMargin)
-        }
-        
-        recentSearchList.snp.makeConstraints { make in
-            make.top.equalTo(recentSearchListTitleLabel.snp.bottom).offset(Metric.List.topMargin)
-            make.left.right.equalToSuperview().inset(Metric.List.horizontalMargin)
-            make.height.equalTo(45 * 5)
-        }
-        
-        tagTitleLabel.snp.makeConstraints { make in
-            make.top.equalTo(recentSearchList.snp.bottom).offset(Metric.TagTitle.topMargin)
-            make.left.equalToSuperview().inset(Metric.TagTitle.leftMargin)
-        }
-        
-        tagCollectionView.snp.makeConstraints { make in
-            make.top.equalTo(tagTitleLabel.snp.bottom).offset(Metric.Tags.topMagin)
-            make.left.right.equalToSuperview().inset(Metric.Tags.horizontalMargin)
-            make.bottom.lessThanOrEqualToSuperview().offset(14)
-        }
     }
 }
